triggerpurviewscan/requirements.txt
azure-functions
azure-identity
azure-purview-scanning
azure-purview-administration
azure-core
